+++ VoxReader/ChunkFactory.cs
@@ -28,13 +28,15 @@ namespace VoxReader
                     return new GroupNodeChunk(data);
                 case ChunkType.ShapeNode:
                     return new ShapeNodeChunk(data);
+                case ChunkType.Note:
+                    return new NoteChunk(data);
+                case ChunkType.IndexMap:
+                    return new IndexMapChunk(data);
                 case ChunkType.MaterialOld:
                 case ChunkType.MaterialNew:
                 case ChunkType.Layer:
                 case ChunkType.Object:
                 case ChunkType.Camera:
-                case ChunkType.Note:
-                case ChunkType.IndexMap:
                     return new Chunk(data);
                 default:
                     throw new ArgumentOutOfRangeException();